requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
beautifulsoup4==4.12.2
billiard==4.1.0
bs4==0.0.1
celery==5.3.4
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Deprecated==1.2.14
Django==4.0.10
django-mongoengine==0.5.4
djangorestframework==3.14.0
dnspython==2.3.0
idna==3.4
kombu==5.3.2
lxml==4.9.2
mongoengine==0.27.0
packaging==23.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
pymongo==4.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==4.3.4
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
wrapt==1.15.0
